+++ tests/testthat/helper.R
@@ -45,12 +45,12 @@ expect_not_identical_graphs <- function(g1, g2, ...) {
 }
 
 expect_snapshot_igraph_error <- function(x, ...) {
-  expect_snapshot(
-    x,
+  inject(expect_snapshot(
+    {{ x }},
     error = TRUE,
     transform = function(y) {
       gsub(":(\\d+)", ":xx", y)
     },
     ...
-  )
+  ))
 }